# optimized modulus in hashToFr