An optimized version of `hashToFr` that doesn't require going through the `big.Int` shebang.