@gballet 2021-07-01
A first run of the benchmarks on my laptop doesn't seem to bring much improvement:

master:
```
(master)> go test -tags bignum_kilic . -run none -bench Bench
goos: freebsd
goarch: amd64
pkg: github.com/gballet/go-verkle
cpu: Intel(R) Core(TM) i7-8550U CPU @ 1.80GHz
BenchmarkProofCalculation-8    	       7	 153526089 ns/op	 3145864 B/op	   47294 allocs/op
BenchmarkProofVerification-8   	     530	   2317685 ns/op	   73384 B/op	     563 allocs/op
BenchmarkCommitLeaves/insert/leaves/1000/width/10-8         	       4	 313753670 ns/op	87145152 B/op	  148946 allocs/op
BenchmarkCommitLeaves/insertOrdered/leaves/1000/width/10-8  	       3	 340776185 ns/op	89567418 B/op	  166664 allocs/op
BenchmarkCommitLeaves/insert/leaves/10000/width/10-8        	       1	3765796378 ns/op	835007336 B/op	 1722117 allocs/op
BenchmarkCommitLeaves/insertOrdered/leaves/10000/width/10-8 	       1	3616759060 ns/op	845955392 B/op	 1804988 allocs/op
BenchmarkCommitLeaves/insert/leaves/1000/width/8-8          	       4	 330652026 ns/op	38192516 B/op	  174374 allocs/op
BenchmarkCommitLeaves/insertOrdered/leaves/1000/width/8-8   	       3	 363885650 ns/op	40452781 B/op	  190997 allocs/op
BenchmarkCommitLeaves/insert/leaves/10000/width/8-8         	       1	2963268291 ns/op	303913416 B/op	 1273593 allocs/op
BenchmarkCommitLeaves/insertOrdered/leaves/10000/width/8-8  	       1	3240214319 ns/op	313068344 B/op	 1344545 allocs/op
BenchmarkCommitFullNode/width/10-8                          	       5	 219208747 ns/op	70336467 B/op	  102535 allocs/op
BenchmarkCommitFullNode/width/8-8                           	      21	  56831579 ns/op	 6599444 B/op	   25753 allocs/op
BenchmarkModifyLeaves-8                                     	       1	12516311186 ns/op	991786440 B/op	 6796385 allocs/op
PASS
ok  	github.com/gballet/go-verkle	244.033s
```
this PR:
```
(hashfr-fast)> go test -tags bignum_kilic . -run none -bench Bench
goos: freebsd
goarch: amd64
pkg: github.com/gballet/go-verkle
cpu: Intel(R) Core(TM) i7-8550U CPU @ 1.80GHz
BenchmarkProofCalculation-8    	       7	 151823501 ns/op	 2730664 B/op	   40227 allocs/op
BenchmarkProofVerification-8   	     524	   2332991 ns/op	   73096 B/op	     557 allocs/op
BenchmarkCommitLeaves/insert/leaves/1000/width/10-8         	       4	 316554238 ns/op	88535456 B/op	  147348 allocs/op
BenchmarkCommitLeaves/insertOrdered/leaves/1000/width/10-8  	       3	 373335669 ns/op	91133888 B/op	  166013 allocs/op
BenchmarkCommitLeaves/insert/leaves/10000/width/10-8        	       1	3541300189 ns/op	832663776 B/op	 1677551 allocs/op
BenchmarkCommitLeaves/insertOrdered/leaves/10000/width/10-8 	       1	3671885967 ns/op	842921160 B/op	 1755275 allocs/op
BenchmarkCommitLeaves/insert/leaves/1000/width/8-8          	       4	 332598580 ns/op	38222064 B/op	  170660 allocs/op
BenchmarkCommitLeaves/insertOrdered/leaves/1000/width/8-8   	       3	 430298035 ns/op	40533922 B/op	  187340 allocs/op
BenchmarkCommitLeaves/insert/leaves/10000/width/8-8         	       1	3019925469 ns/op	302519952 B/op	 1234108 allocs/op
BenchmarkCommitLeaves/insertOrdered/leaves/10000/width/8-8  	       1	3112502382 ns/op	311511560 B/op	 1303068 allocs/op
BenchmarkCommitFullNode/width/10-8                          	       5	 212641849 ns/op	70139859 B/op	   98439 allocs/op
BenchmarkCommitFullNode/width/8-8                           	      21	  52969637 ns/op	 6550293 B/op	   24729 allocs/op
BenchmarkModifyLeaves-8                                     	       1	12545886955 ns/op	973719744 B/op	 6391371 allocs/op
PASS
ok  	github.com/gballet/go-verkle	244.167s
```