@gballet 2021-07-01
Running in on an AWS x5.large, the gains are more marked:

master:
```
goos: linux
goarch: amd64
pkg: github.com/gballet/go-verkle
cpu: Intel(R) Xeon(R) Platinum 8175M CPU @ 2.50GHz
BenchmarkProofCalculation-4    	       9	 119828718 ns/op	 3159024 B/op	   47579 allocs/op
BenchmarkProofVerification-4   	     676	   1790521 ns/op	   73384 B/op	     563 allocs/op
BenchmarkCommitLeaves/insert/leaves/1000/width/10-4         	       5	 244006858 ns/op	87949697 B/op	  150699 allocs/op
BenchmarkCommitLeaves/insertOrdered/leaves/1000/width/10-4  	       4	 267421160 ns/op	90470608 B/op	  169190 allocs/op
BenchmarkCommitLeaves/insert/leaves/10000/width/10-4        	       1	2639188304 ns/op	834225064 B/op	 1720883 allocs/op
BenchmarkCommitLeaves/insertOrdered/leaves/10000/width/10-4 	       1	2806947636 ns/op	844845280 B/op	 1802626 allocs/op
BenchmarkCommitLeaves/insert/leaves/1000/width/8-4          	       4	 254489847 ns/op	38253672 B/op	  174574 allocs/op
BenchmarkCommitLeaves/insertOrdered/leaves/1000/width/8-4   	       4	 279418100 ns/op	40528266 B/op	  191330 allocs/op
BenchmarkCommitLeaves/insert/leaves/10000/width/8-4         	       1	2285747460 ns/op	304501192 B/op	 1276876 allocs/op
BenchmarkCommitLeaves/insertOrdered/leaves/10000/width/8-4  	       1	2396715266 ns/op	313690968 B/op	 1348501 allocs/op
BenchmarkCommitFullNode/width/10-4                          	       7	 161655760 ns/op	70336469 B/op	  102535 allocs/op
BenchmarkCommitFullNode/width/8-4                           	      30	  38834775 ns/op	 6599442 B/op	   25753 allocs/op
BenchmarkModifyLeaves-4                                     	       1	10464597561 ns/op	994717400 B/op	 6804008 allocs/op
PASS
ok  	github.com/gballet/go-verkle	198.060s
```

This PR:
```
goos: linux
goarch: amd64
pkg: github.com/gballet/go-verkle
cpu: Intel(R) Xeon(R) Platinum 8175M CPU @ 2.50GHz
BenchmarkProofCalculation-4    	       9	 118977268 ns/op	 2728312 B/op	   40179 allocs/op
BenchmarkProofVerification-4   	     673	   1764442 ns/op	   73096 B/op	     557 allocs/op
BenchmarkCommitLeaves/insert/leaves/1000/width/10-4         	       5	 244374198 ns/op	88431158 B/op	  147975 allocs/op
BenchmarkCommitLeaves/insertOrdered/leaves/1000/width/10-4  	       4	 269565498 ns/op	90992576 B/op	  166384 allocs/op
BenchmarkCommitLeaves/insert/leaves/10000/width/10-4        	       1	2626937048 ns/op	832544016 B/op	 1677367 allocs/op
BenchmarkCommitLeaves/insertOrdered/leaves/10000/width/10-4 	       1	2797238624 ns/op	842839016 B/op	 1755355 allocs/op
BenchmarkCommitLeaves/insert/leaves/1000/width/8-4          	       4	 255269312 ns/op	38011214 B/op	  169882 allocs/op
BenchmarkCommitLeaves/insertOrdered/leaves/1000/width/8-4   	       4	 278445845 ns/op	40193708 B/op	  185665 allocs/op
BenchmarkCommitLeaves/insert/leaves/10000/width/8-4         	       1	2282395358 ns/op	301746176 B/op	 1228477 allocs/op
BenchmarkCommitLeaves/insertOrdered/leaves/10000/width/8-4  	       1	2395500513 ns/op	310773576 B/op	 1297678 allocs/op
BenchmarkCommitFullNode/width/10-4                          	       7	 161509285 ns/op	70139858 B/op	   98439 allocs/op
BenchmarkCommitFullNode/width/8-4                           	      31	  38752496 ns/op	 6550296 B/op	   24729 allocs/op
BenchmarkModifyLeaves-4                                     	       1	9828710262 ns/op	971894896 B/op	 6385651 allocs/op
PASS
ok  	github.com/gballet/go-verkle	197.477s
```